linux_tfod1_requirements.txt
absl-py==1.3.0
aioice==0.7.6
aiortc==1.3.2
altair==4.2.0
anyio==3.5.0
apache-beam==2.42.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==21.4.0
av==9.2.0
babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.5
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.4
cloudpickle==2.1.0
colorama==0.4.5
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.5
crcmod==1.7
cryptography==37.0.1
cvzone==1.5.6
cycler==0.11.0
cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dm-tree==0.1.7
dnspython==2.2.1
docopt==0.6.2
easyocr==1.6.2
entrypoints==0.4
etils==0.8.0
executing==1.1.1
fastavro==1.6.1
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.9
gitpython==3.1.29
google-api-python-client==2.65.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.4
imageio==2.22.4
immutabledict==2.2.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.0
ipython==8.5.0
ipython_genutils==0.2.0
jedi==0.18.1
jinja2==3.0.3
json5==0.9.6
jsonschema==4.16.0
jupyter-client==7.4.2
jupyter_core==4.11.1
jupyter_server==1.18.1
jupyterlab==3.4.4
jupyterlab_pygments==0.1.2
jupyterlab_server==2.15.2
keras==2.10.0
keras-preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lvis==0.5.3
lxml==4.8.0
markdown==3.4.1
markupsafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mediapipe==0.9.0
mistune==0.8.4
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==2.8.8
ninja==1.11.1
notebook==6.4.12
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.1
#object-detection==0.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.5.2.52
opt-einsum==3.3.0
orjson==3.8.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==9.2.0
pip==22.3.1
ply==3.11
portalocker==2.6.0
prometheus_client==0.14.1
promise==2.3
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==3.19.6
psutil==5.9.2
pure_eval==0.2.2
py-cpuinfo==8.0.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
pycocotools==2.0.5
pycparser==2.21
pydeck==0.8.0
pydot==1.4.2
pyee==9.0.4
pygments==2.13.0
pylibsrtp==0.7.1
pymongo==3.12.3
pympler==1.0.1
pyopenssl==22.0.0
pyparsing==2.4.7
pyqt5-sip==12.11.0
pyrsistent==0.18.0
pysocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywavelets==1.4.1
pywin32==304
pywinpty==2.0.2
pyyaml==5.3.1
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
sacrebleu==2.2.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
semver==2.13.0
send2trash==1.8.0
sentencepiece==0.1.97
setuptools==63.4.1
shapely==1.8.5.post1
sip==6.6.2
six==1.16.0
slim==0.1
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
split-folders==0.5.1
stack-data==0.5.1
streamlit==1.14.1
streamlit-webrtc==0.44.0
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-addons==0.18.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-io==0.27.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.10.0
tensorflow-model-optimization==0.7.3
tensorflow-text==2.10.0
termcolor==2.0.1
terminado==0.13.1
testpath==0.6.0
text-unidecode==1.3
tf-models-official==2.10.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2022.10.10
toml==0.10.2
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typeguard==2.13.3
typing-extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
werkzeug==2.2.2
wget==3.2
wheel==0.37.1
win_inet_pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
zipp==3.9.0
zstandard==0.18.0
zstd==1.5.2